feat: rename graphlib to @dagrejs/graphlib and merge types of @dagrejs/graphlib and dagre
