+++ types/dagre-d3/dagre-d3-tests.ts
@@ -1,20 +1,13 @@
-const gDagre = new dagreD3.graphlib.Graph();
-const graph = gDagre.graph();
+const graph = new dagreD3.graphlib.Graph();
 
 // has graph methods from dagre.d.ts
 graph.setNode("a", {});
-const num: number = 251 + graph.height + graph.width;
-const predecessors: { [vertex: string]: string[] } = {};
-const successors: { [vertex: string]: string[] } = {};
-
-predecessors["a"] = graph.predecessors("a");
-successors["a"] = graph.successors("a");
 graph.transition = (selection: d3.Selection<any>) => {
     return d3.transition();
 };
 
 const render = new dagreD3.render();
 const svg = d3.select("svg");
-render.arrows()["arrowType"] = (parent: d3.Selection<any>, id: string, edge: dagre.Edge, type: string) => {};
-render.shapes()["shapeName"] = (parent: d3.Selection<any>, bbox: any, node: dagre.Node) => {};
+render.arrows()['arrowType'] = (parent: d3.Selection<any>, id: string, edge: dagre.graphlib.Edge, type: string) => {};
+render.shapes()['shapeName'] = (parent: d3.Selection<any>, bbox: any, node: dagre.graphlib.Node) => {};
 render(svg, graph);
