+++ types/dagre-d3/index.d.ts
@@ -11,27 +11,18 @@ export as namespace dagreD3;
 
 export * from "dagre";
 export const render: { new(): Render };
-export const intersect: { [shapeName: string]: (node: dagre.Node, points: Array<{}>, point: any) => void };
+export const intersect: { [shapeName: string]: (node: dagre.graphlib.Node, points: Array<{}>, point: any) => void };
 
-declare module "dagre" {
-    namespace graphlib {
-        // coupled with dagre.d.ts' Graph
-        // a lot of these methods come from graphlib.core.js
-        interface Graph {
-            graph(): Graph;
-            height: number;
-            predecessors(id: string): string[];
-            successors(id: string): string[];
-            // see source of http://cpettitt.github.io/project/dagre-d3/latest/demo/interactive-demo.html
-            transition?(selection: d3.Selection<any>): d3.Transition<any>;
-            width: number;
-        }
+declare module "@dagrejs/graphlib" {
+    interface Graph {
+        // see source of http://cpettitt.github.io/project/dagre-d3/latest/demo/interactive-demo.html
+        transition?(selection: d3.Selection<any>): d3.Transition<any>;
     }
 }
 
 export interface Render {
     // see http://cpettitt.github.io/project/dagre-d3/latest/demo/user-defined.html for example usage
-    arrows(): { [arrowStyleName: string]: (parent: d3.Selection<any>, id: string, edge: dagre.Edge, type: string) => void };
+    arrows(): { [arrowStyleName: string]: (parent: d3.Selection<any>, id: string, edge: dagre.graphlib.Edge, type: string) => void };
     (selection: d3.Selection<any>, g: dagre.graphlib.Graph): void;
-    shapes(): { [shapeStyleName: string]: (parent: d3.Selection<any>, bbox: any, node: dagre.Node) => void };
+    shapes(): { [shapeStyleName: string]: (parent: d3.Selection<any>, bbox: any, node: dagre.graphlib.Node) => void };
 }
