+++ types/dagre/dagre-tests.ts
@@ -30,10 +30,10 @@ dagre.graphlib.alg.components(gDagre);
 dagre.graphlib.alg.dijkstra(gDagre, 'a', (edge) => 5);
 dagre.graphlib.alg.dijkstraAll(gDagre);
 dagre.graphlib.alg.findCycles(gDagre);
-dagre.graphlib.alg.floydWarchall(gDagre);
+dagre.graphlib.alg.floydWarshall(gDagre);
 dagre.graphlib.alg.isAcyclic(gDagre);
 dagre.graphlib.alg.postorder(gDagre, 'a');
 dagre.graphlib.alg.preorder(gDagre, ['b', 'c']);
 dagre.graphlib.alg.prim(gDagre);
-dagre.graphlib.alg.tarjam(gDagre);
+dagre.graphlib.alg.tarjan(gDagre);
 dagre.graphlib.alg.topsort(gDagre);
