+++ types/dagre/tsconfig.json
@@ -14,6 +14,9 @@
         ],
         "types": [],
         "noEmit": true,
+        "paths": {
+            "@dagrejs/graphlib": ["dagrejs__graphlib"]
+        },
         "forceConsistentCasingInFileNames": true
     },
     "files": [
