+++ types/dagrejs__graphlib/tsconfig.json
@@ -6,7 +6,7 @@
         ],
         "noImplicitAny": true,
         "noImplicitThis": true,
-        "strictNullChecks": false,
+        "strictNullChecks": true,
         "strictFunctionTypes": true,
         "baseUrl": "../",
         "typeRoots": [
@@ -14,10 +14,13 @@
         ],
         "types": [],
         "noEmit": true,
+        "paths": {
+            "@dagrejs/graphlib": ["dagrejs__graphlib"]
+        },
         "forceConsistentCasingInFileNames": true
     },
     "files": [
         "index.d.ts",
-        "graphlib-tests.ts"
+        "dagrejs__graphlib-tests.ts"
     ]
 }
